+++ src/main/java/com/aaronpmaus/jProt/executables/JProtMetrics.java
@@ -28,14 +28,9 @@ import java.util.Date;
 *      -a, --angular-distance
 *          Calculate and print the angular distance.
 *      --mol1-f fname
-*          This file must either be a PDB file or a csv file. If it is a
-*          PDB file it must end in the .pdb extension and conform to the
-*          PDB File Format Version 3.30. otherwise, the file should
-*          contain the distance matrix for If the file provided is csv,
-*          it must end in the .csv extension. The first row must contain
-*          the residue IDs and the rest of the file contains all pairwise
-*          distances between those residues.  Each row contains that
-*          residues distances to every other residue.
+*          This file must either be a PDB file. If it is a PDB file it
+*          must end in the .pdb extension and conform to the PDB File
+*          Format Version 3.30.
 *      --mol2-f fname
 *          The file for molecule two. The format is the same as mol1-f.
 *      --ls, --local-similarity
@@ -63,6 +58,10 @@ import java.util.Date;
 *      --gdt-plot
 *          Global Distance Test: Operates as --gdt but with thresholds
 *          {0.5, 1.0, 1.5, ..., 9.5, 10.0}
+*      --chimera
+*          Print out a chimera script to color the regions of similarity
+*          found by either local similarity or gdt. By default, a pymol
+*          script is printed. This changes that behavior.
 * </code>
 * </pre>
 *
@@ -74,8 +73,9 @@ public class JProtMetrics{
   private static boolean runLocalSimilarity = false;
   private static boolean runGDT = false;
   private static boolean printGDTPlotData = false;
+  private static boolean runGDTHA = false;
   private static boolean usePDBs = false;
-  private static boolean useCSVs = false;
+  private static boolean printChimera = false;
   private static double localSimilarityThreshold = 1.0;
   private static double[] gdtThresholds;
   private static String mol1FilePath;
@@ -84,6 +84,7 @@ public class JProtMetrics{
   private static boolean diffFileProvided = false;
   private static boolean mol1FileProvided = false;
   private static boolean mol2FileProvided = false;
+  private static boolean runningGDTHA = false;
   private static Metrics theTool;
   /**
   * Runs the metrics of this program.
@@ -113,14 +114,11 @@ public class JProtMetrics{
         runGDT = true;
       }
       if(args.contains("--gdt-ha")){
-        runGDT = true;
-        gdtThresholds[0] = 0.5;
-        gdtThresholds[1] = 1.0;
-        gdtThresholds[2] = 2.0;
-        gdtThresholds[3] = 4.0;
+        runGDTHA = true;
       }
       if(args.contains("--gdt-plot")){
         runGDT = true;
+        runGDTHA = false;
         printGDTPlotData = true;
         gdtThresholds = new double[20];
         for(int i = 0; i < 20; i++){
@@ -142,6 +140,9 @@ public class JProtMetrics{
         mol2FilePath = args.getValue("--mol2-f");
         mol2FileProvided = true;
       }
+      if(args.contains("--chimera")){
+        printChimera = true;
+      }
     }
 
     try {
@@ -158,11 +159,8 @@ public class JProtMetrics{
 
         if(mol1Ext.equals("pdb") && mol2Ext.equals("pdb")){
           usePDBs = true;
-        } else if(mol1Ext.equals("csv") && mol2Ext.equals("csv")){
-          useCSVs = true;
         } else {
-          System.out.println("Both files must either be PDBs with extension .pdb"
-            + " or CSVs with extension .csv");
+          System.out.println("Both files must either be PDBs with extension .pdb");
           System.out.println("You provided files: \n" + mol1FileName + "\n" + mol2FileName);
           System.exit(1);
         }
@@ -174,8 +172,6 @@ public class JProtMetrics{
           Protein prot2 = new PDBFileIO().readInPDBFile(new FileInputStream(mol2FileName),mol2Base);
           //Protein prot2 = pdb.readInPDBFile(mol2FileName);
           theTool = new Metrics(prot1, prot2);
-        } else if(useCSVs){
-          theTool = new Metrics(mol1FileName, mol2FileName);
         }
       } else {
         System.out.println("You must provide the two protein data files.");
@@ -184,6 +180,14 @@ public class JProtMetrics{
       if(runAngularDistance) angularDistance();
       if(runLocalSimilarity) localSimilarity(localSimilarityThreshold);
       if(runGDT) globalDistanceTest(gdtThresholds);
+      gdtThresholds[0] = 0.5;
+      gdtThresholds[1] = 1.0;
+      gdtThresholds[2] = 2.0;
+      gdtThresholds[3] = 4.0;
+      if(runGDTHA) {
+        runningGDTHA = true;
+        globalDistanceTest(gdtThresholds);
+      }
     } catch (FileNotFoundException e){
       System.out.println("Could not open required files. Check for existence.");
       System.exit(1);
@@ -218,13 +222,20 @@ public class JProtMetrics{
     ArrayList<UndirectedGraph<Integer>> localSimilarityRegions;
     localSimilarityRegions = theTool.getLocalSimilarityRegions(threshold);
 
-    ArrayList<String> pymolScript = theTool.getPymolColoringScript(localSimilarityRegions);
+    ArrayList<String> script;
 
-    System.out.println("\n#Pymol Script:");
-    for(String cmd: pymolScript){
+    if(printChimera){
+      script = theTool.getChimeraColoringScript(localSimilarityRegions);
+      System.out.println("\n#Chimera Script:");
+    } else {
+      script = theTool.getPymolColoringScript(localSimilarityRegions);
+      System.out.println("\n#Pymol Script:");
+    }
+
+    for(String cmd: script){
       System.out.println(cmd);
     }
-    System.out.println("#End of Pymol Script\n");
+    System.out.println("#End of Script\n");
 
     double[][] globalDistanceTest = theTool.getGlobalDistanceTestScore(localSimilarityRegions);
     String cliquesStr = "Cliques:";
@@ -263,13 +274,20 @@ public class JProtMetrics{
     int numThresholds = thresholds.length;
     globalDistanceRegions = theTool.getGlobalDistanceRegions(thresholds);
 
-    ArrayList<String> pymolScript = theTool.getPymolColoringScript(globalDistanceRegions);
+    ArrayList<String> script;
 
-    System.out.println("\n#Pymol Script:");
-    for(String cmd: pymolScript){
+    if(printChimera){
+      script = theTool.getChimeraColoringScript(globalDistanceRegions);
+      System.out.println("\n#Chimera Script:");
+    } else {
+      script = theTool.getPymolColoringScript(globalDistanceRegions);
+      System.out.println("\n#Pymol Script:");
+    }
+
+    for(String cmd: script){
       System.out.println(cmd);
     }
-    System.out.println("#End of Pymol Script\n");
+    System.out.println("#End of Script\n");
 
     double[][] globalDistanceTest = theTool.getGlobalDistanceTestScore(globalDistanceRegions);
     System.out.println("Global Distance Test:");
@@ -301,13 +319,17 @@ public class JProtMetrics{
       double mcdtScore_ha = ( half + one + two + four ) / 4.0;
       double mcdtScore = ( one + two + four + eight) / 4.0;
       System.out.println();
-      System.out.printf("MCDT-HA Score: %.4f\n", mcdtScore_ha);
-      System.out.printf("MCDT Score: %.4f\n", mcdtScore);
+      System.out.printf("RoS-GDT-HA Score: %.4f\n", mcdtScore_ha);
+      System.out.printf("RoS-GDT Score: %.4f\n", mcdtScore);
     } else {
       // The last row in the array holds the averages. If there are 4 thresholds,
       // rows 0-3 hold the number and percents of residues for each threshold. Row
       // 4 holds the averages.
-      System.out.printf("MCDT Score: %.4f\n",globalDistanceTest[numThresholds][1]*100);
+      if(runningGDTHA){
+        System.out.printf("RoS-GDT-HA Score: %.4f\n",globalDistanceTest[numThresholds][1]*100);
+      } else {
+        System.out.printf("RoS-GDT Score: %.4f\n",globalDistanceTest[numThresholds][1]*100);
+      }
     }
     long end = new Date().getTime();
     System.out.println("Total Time for Global Distance Test: " + (end - start) + " milleseconds.");